shared/sec: change log label in cmd parser
In cmd_parser.c, two of labels are used in log messages, `APP` and
`SPP_COMMAND_PROC`. First one is almost no meaning and second one is
not suitable considering the feature of this module. In addition, no
need to use two labels.

This update is to change both of labels to `WK_CMD_PARSER`.

Signed-off-by: Yasufumi Ogawa <ogawa.yasufumi@lab.ntt.co.jp>