bin: make template config as default
Config file of startup script `bin/env.sh` is under management of git,
but it is not convenient for developers if this file is changed and
asked to `git add` even if subtle changes. This update is to add
template config under the management of git and to use customized
config which is not included under git.

The startup script expects to find customized `bin/config.sh`, or create
it by copying template `bin/sample/config.sh` if it does not exist.

Signed-off-by: Yasufumi Ogawa <ogawa.yasufumi@lab.ntt.co.jp>