shared/sec: rename defines for length of str
This update is to rename `SPP_MIN_STR_LEN` to `STR_LEN_SHORT` because it
is used not for the minimum length but for just comparably short length
of strings. It also renames `SPP_NAME_STR_LEN` to `STR_LEN_NAME` used
for size of names of component or interfaces.

Signed-off-by: Yasufumi Ogawa <yasufum.o@gmail.com>