shared/sec: refactor setup cls table stat in JSON
This update is to refactor setting up a list of entries in classifier
table for `status` command. Names of function for setting up JSON are
revised, but still hard to understand the features.
append_classifier_table() for setting up the entries calls several
nested functions as in the following order.

  append_classifier_table()  // This function.
    -> add_classifier_table()  // Wrapper and doesn't almost nothing.
      -> add_classifier_table_val()  // Setup data and call iterator.
        -> iterate_adding_mac_entry()

Signed-off-by: Yasufumi Ogawa <yasufum.o@gmail.com>