shared/sec: change order of args of JSON fmtters
Formatter functions defined in `json_helper.h` takes three args, a pair
of name and value of JSON and a string buffer as a placeholder. But the
order of args `func(name, buf, value)` is not well. This update is to
move `buf` at first as similar to other C functions such as sprintf().

Signed-off-by: Yasufumi Ogawa <yasufum.o@gmail.com>