shared/sec: move JSON formatter to shard/secondary
Principle JSON formatters, such as append_json_int_value(), can be used
from all of SPP secondary processes. This update is to move files
`json_helper.*` and `string_buffer.*` to the parent directory
`shared/secondary`. Definition of return codes of `SPP_RET_OK` and
`SPP_RET_NG` is also moved to the directory for the change.

Signed-off-by: Yasufumi Ogawa <yasufum.o@gmail.com>