shared/sec: move getter funcs for worker status
Following getter functinos for spp_vf and spp_mirror are defined in
shared directory and worker specific tasks are compiled by using
`#ifdef SPP_VF_MODULE` and `#ifdef SPP_MIRROR_MODULE` selectively.
  * get_forwarder_status()
  * get_classifier_status()
  * get_mirror_status()

However, it tends to be messed up because common and specific functions
are mixed up. It should be separated cleanly. This update is to fix the
issue by reorganizing them and remove ifdef switches.

Signed-off-by: Yasufumi Ogawa <yasufum.o@gmail.com>