controller: move vf help msg to SppVf.help
SPP CLI is implemented with python's cmd library and help message is
defined as a comment of `do_*` method. However, total number of lines
tends to be a large number if the number of command is increased.

To avoid such a situation, this update is to move comment of `do_vf`
to `SppVf.help` to reduce the number of lines.

Signed-off-by: Yasufumi Ogawa <yasufum.o@gmail.com>