shared/sec: revise enum spp_port_rxtx
This update is to revise the name of enum `spp_port_rxtx` to
`sppwk_port_dir` because it is used to specify which of direction
packets are forwarded on a port. Term `rxtx` is used for a set of RX and
TX, so it should not be used for specifying the direction. This update
is also rename a member of structs using this enum from `rxtx` to `dir`.

Some error are occured while compiling spp_pcap for this update. It is
fixed by next pathces.

Signed-off-by: Yasufumi Ogawa <yasufum.o@gmail.com>