shared/sec: rename structs of port ability
In spp_vf, attribute of port such as direction, operation functions or
VLAN tag info as port ability. However, names of structures are not
appropriate, for example `spp_port_ability` is owned by
`port_ability_mng_info` to contain a set of abilities and it is owned
by `port_ability_port_mng_info` to contain other attributes. It is so
confusing and not describing the features.

This update is to fix the issue, to `port_abl_mng_info` has
`port_abl_info`.

Signed-off-by: Yasufumi Ogawa <yasufum.o@gmail.com>