shared/sec: remove function spp_vf_add_vhost_pmd
Spp_vf and spp_mirror use spp_vf_add_vhost_pmd() for assigning vhost
port, but this feature is the same as add_vhost_pmd(). This update is to
remove former one. For this update, remove `vhost_client` which is a
member of struct `startup_param` and no longer needed.

Signed-off-by: Yasufumi Ogawa <yasufum.o@gmail.com>