shared/sec: remove worker process type
Three of SPP worker process types are defined as an enum for vf,
mirror and none. It is only used as index for getting name of types of
string value from an array.

  enum sppwk_proc_type {
       WK_PROC_TYPE_NONE,
       WK_PROC_TYPE_VF,
       WK_PROC_TYPE_MIRROR,
  }

However, it is menaingless because type for vf is not used in
spp_mirror, or vise versa. In addition, type none is not used from
anywhere. It should be simply replaced with fixed string values in each
of processes because it is not needed to be shared among processes.

This update is to replace this enum and array of string with
`#define WK_PROC_TYPE "vf"` and `#define WK_PROC_TYPE "mirror"` in each
of processes independently.

Signed-off-by: Yasufumi Ogawa <yasufum.o@gmail.com>