spp_nfv: remove global variable client_id
`client_id` is defined as a global variable in `params.h`, but it is
better to use get_client_id() instead. This update is to replace it.
This update is also to make a change for formatter function
`get_sec_stats_json()` which takes client ID as uint16_t, but
get_client_id() returns the ID as int.

Signed-off-by: Yasufumi Ogawa <yasufum.o@gmail.com>