shared/sec: Add RX and TX burst without VLAN
In SPP worker, original version of RX and TX burst functions are
implemented in which adding or deleting VLAN tag are added to
rte_eth_rx_burst() and rte_eth_tx_burst(). However, it is also called
from processes which does not VLAN features.

This update is to add functions without VLAN features in addition to
them for supporting both of usecases.

  * sppwk_eth_rx_burst()
  * sppwk_eth_tx_burst()
  * sppwk_eth_vlan_rx_burst()
  * sppwk_eth_vlan_tx_burst()

Signed-off-by: Yasufumi Ogawa <yasufum.o@gmail.com>