shared: add parsing ethdev name
This update is to add parser parse_dev_name() for the name of ethdev,
such as `eth_vhost1` or `net_ring1`. By using this function,
port type `VHOST` and ID `1` can be extracted from `eth_vhost1`.

This parser is intended to be used in initialization of ports in
secondary process in which all shared ports managed by spp_primary are
initialized as PHY incorrectly even if it is not PHY type. It is
because secondary expects the type of ports in spp_primary is always
PHY. However, port type might be other than PHY if spp_primary is
launched with `--vdev` option.

By using this parser, port type and ID can be inspected in secondary.

Signed-off-by: Yasufumi Ogawa <yasufum.o@gmail.com>