spp_nfv: parse port type and ID while init ports
Spp_nfv initialize shared ports managed by spp_primary as PHY type.
However, it might be other types if spp_primary is launched with
`--vdev` option, and it should be initialized with correct port type
and ID.

This update is to fix the issue by using parse_dev_name() which is added
in the previous patch.

Signed-off-by: Yasufumi Ogawa <yasufum.o@gmail.com>