shared/sec: remove nof_rings from iface_info
Struct `iface_info` is for containing the number of ports and port info
itself for each of port types. However, members of the number of ports
are scarcely used, only for initializing ring latency feature.

As the result, three of `nof_phys`, `nof_vhosts` and `nof_rings` are not
required. Former two are not referred from anywhere, and `nof_rings` can
be counted with RTE libs. This patch is to remove it.

Signed-off-by: Yasufumi Ogawa <yasufum.o@gmail.com>