shared/sec: move data types from cmd_utils
In `cmd_utils.h`, the rest of enums, structs or functions which are not
included in other specific groups, such as parser or formatter, are
defined temporarily. The term `utils` means as same as `others`.

This update is to move enums and structs roughly grouped in
`cmd_utils.h` to `data_types.h`. This file is to define common data
types among SPP worker processes, but specific data for spp_vf are
still remained actually. Specific data types are moved to outside of
shared directory in future patches.

Signed-off-by: Yasufumi Ogawa <yasufum.o@gmail.com>