spp_vf: remove nouse wrapper function
add_classifier_table() is called for adding contents of classifier table
to the result of `status` command. This function calls several functions
as following.

    add_classifier_table()
      -> _add_classifier_table()  /* Wrapper doesn't almost nothing. */
        -> add_classifier_table_val()
          -> iterate_adding_mac_entries()

However, this wrapper function does not anything actually, and the names
of functions are not appropriate considering its behaviours. This update
is to remove nouse wrapper function.

Signed-off-by: Yasufumi Ogawa <yasufum.o@gmail.com>