cli: fix used sec ID is not updated
Shared variable `spp_common.used_sec_ids` was introduced to avoid
that two secondary processes are launched with duplicated sec IDs.
However, it is not updated if secondary process is terminated, and
its ID cannot be used again although the ID is released actually.

To fix the issue, change to get used sec IDs everytime it is referred
and remove nouse `spp_common.used_sec_ids`.

Fixes: 6da5038472d9 ("controller: avoid launch with duplicated IDs")

Signed-off-by: Yasufumi Ogawa <yasufum.o@gmail.com>