cli: revise template of pri launch command
For generating command line options for `pri; launch` dynamically, SPP
CLI has a template of command line in `commands/pri.py`. Labels in the
template are replaced with actual params while running command
completion.

Style of labels is mixed as `{keyword}` and `__XXX__` for replacing them
in several steps. Replace each of `__XXX__` style labels step by step,
then `{keyword}` labels with format() finally. However, it is complex
and better to unify all labels with `{keyword}` and replace at once.

This update is to unify label style to `{keyword}` and replace them at
one time. It also remove _setup_launch_template() for replacing
`__XXX__` because it is no longer required.

Signed-off-by: Yasufumi Ogawa <yasufum.o@gmail.com>