cli: revise composing dot script
SppTopo uses old style formatter, such as '"%s %s" % (foo, var)', for
composing dot script. It could be complex if several placeholders are
included in a string. It is better to replace latest style using
format() method.

This update is to replace all formatters for using format(), and use
labels in placeholders to be more understandable.

Signed-off-by: Yasufumi Ogawa <yasufum.o@gmail.com>