cli: add to get sec ID and procs to SppCtlClient
Shell class has get_sec_ids() for retrieving secondary IDs of given
type by requesting all processes information to spp-ctl. However, it is
better to request the information from SppCtlClient which a delegator
for the requesting.

This update is to remove get_sec_ids() from Shell and add the same
method to SppCtlClient.

In addition to above update,  add get_sec_procs() to SppCtlClient to be
able to retrieve other than spp_nfv in topo. By this change, all types
of secondary process can be referred from anywhere in SppTopo so that
there is no need to pass `sec_list` as an argument of member methods.

Supporting other than spp_nfv in topo command is added in the next
patches.

Signed-off-by: Yasufumi Ogawa <yasufum.o@gmail.com>