tools/sppc: add blacklist and whitelist options
EAL option `blacklist` is to exclude PCI devices from under management
of DPDK, and `whitelist` is to include devices in opposition. This
update is to add thesee options to app launcher because app container
tries to have PCI devices, such as `phy:0` or so,  which are already
owned by SPP without `blacklist` or `whitelist`.

Signed-off-by: Yasufumi Ogawa <yasufum.o@gmail.com>