tools/sppc: add suricata build for ubuntu1604
Script for building an image of suricata-4.1.4 on Ubuntu 18.04 was
released as previous patch. This update is to add support for Ubuntu
16.04.

Building suricata-4.1.4 on Ubuntu 16.04 image has some problems and you
cannot complete to compile suricata while docker building. So, you need
to compile suricata by yourself after launched a container of the build
image which is named as `sppc/suricata-ubuntu:16.04` by default.

  $ ./main.py -t suricata --dpdk-branch v18.11 --dist-ver 16.04
  $ docker images
  ...
  sppc/suricata-ubuntu  16.04 ...
  $ docker run -it sppc/suricata-ubuntu:16.04 /bin/bash
  # ./install_suricata.sh
  ...

After compilation is successfully done, logout and create a new docker
image on which suricata is already built with `docker commit` command
and image's container ID.
In this example, new image is named as `sppc/suricata-ubuntu2:16.04`.

  # exit
  $ docker ps -a
  CONTAINER_ID  sppc/suricata-ubuntu:16.04  "/bin/bash"  3 minutes ...
  $ docker commit CONTAINER_ID sppc/suricata-ubuntu2:16.04

Finally, you can run suricata with the new image.

  $ docker run -it sppc/suricata-ubuntu:16.04 /bin/bash
  # suricata --build-info

These steps are described in documents in the next patches.

Signed-off-by: Yasufumi Ogawa <yasufum.o@gmail.com>