shared: add TODO to fix bug in latency stats
Spp_vf and spp_mirror has a feature for measuring latency of ring PMD.
This optional feature is intended to be used by developers and is not
activate by default. It is activated by editing Makefile in src/vf and
src/mirror.

Although this feature is not interested by most of users and does not
prevent users from using SPP, it should be fixed. This update is to add
TODO comments for notifying users to not use this feature until the bug
is fixed.

Signed-off-by: Yasufumi Ogawa <yasufum.o@gmail.com>