spp_primary: add basic forwarding to primary
As vdev has changed to be managed by primary and shared among
secondaries, and this change is not completed yet, creating vdev
from secondary does not work for several devices.

In SPP, primary is a resource manager and secondary is a worker for
creating devices and doing specific tasks using the devices. However,
creating devs from secondary is not doable for above change. So, most
useful vhost cannot be used for the reason.

To remedy the issue, add tasks of creating vdevs and forwarding packets
to primary. Implementation of accepting commands is done in the next
patches.

Signed-off-by: Yasufumi Ogawa <yasufum.o@gmail.com>