tools/sppc: separate building container image
Building container image takes a long time for retrieving several
packages and compile them. This update is to divide each Dockerfile
into installing with apt install and building DPDK and its apps.
In the second step, it uses the image created in the first step.

Files named as `Dockerfile.*.base` are for installing packages, and no
need to run them basically because this built image is uploaded to
docker hub. You just run Dockerfiles which do not have `base` in the
name. You only need to run `Dockerfile.*.base` if you make some changes
in the file, or in offline.

Signed-off-by: Yasufumi Ogawa <yasufum.o@gmail.com>