test: add test for spp_nfv
This patch is to add a unit test code for spp_nfv. Test cases are
defined by following SPP REST API, such as checking status, adding port
or so.

Configurations for the tests are defined as `test/config.ini`. Host IP
or assignment of resouces can be changed in the config.

Unit tests are expected to be run in the `test` dir.

  $ cd /path/to/spp/test
  $ python3 -m unittest  # or add `-v` for showing details

Signed-off-by: Yasufumi Ogawa <yasufum.o@gmail.com>