test: add test for spp_primary
This patch is to add a unit test code for spp_primary. Test cases are
defined by following SPP REST API as similar to spp_nfv.

Unit tests are expected to be run in the `test` dir. Specify the name
of test if you run tests for a process, or all processes without the
name.

  $ cd /path/to/spp/test
  $ python3 -m unittest test_spp_primary  # test only spp_primary
  # or
  $ python3 -m unittest  # test all

Signed-off-by: Yasufumi Ogawa <yasufum.o@gmail.com>