bin: add memif to vdevs of primary
This update is to add memif vdev to spp_primary. The set of IDs of memif
is defined in `bin/config.sh` as following.

  PRI_MEMIF_VDEVS=(0 1)  # IDs of `net_memif`

This memif is attached as master role, and socket file is shared among
all vdevs. As same as vhost socks, it must be deleted before
re-launching, or failed to launch because of socket is already exists.

  $ sudo spp_primary ... \
  --vdev net_memif0,id=0,role=master,socket=/tmp/spp-memif.sock \
  --vdev net_memif1,id=1,role=master,socket=/tmp/spp-memif.sock \
  ...

For client app, it should refer the memif vdev as slave role. It means
simply no need to add `role=master`.

  $ sudo some-dpdk-client-app ... \
  --vdev net_memif0,id=0,socket=/tmp/spp-memif.sock \
  --vdev net_memif1,id=1,socket=/tmp/spp-memif.sock \
  ...

As a note, sock file of memif is created as `/run/memif.sock` as default
if it is not given explicitly. To avoid the implicity, create SPP's sock
file and delete it before re-launch in startup scripts.

Signed-off-by: Yasufumi Ogawa <yasufum.o@gmail.com>