drivers/vhost: add multi process supported vhost
This update is a temporary solution because for vhost PMD can not be
used from secondary processes since DPDK 18.11.

This patch is to add a dedicated vhost PMD other than DPDK's for SPP
secondary processes at the moment. This vhost PMD is derived from DPDK's
implementation, but simplified and features are limited which are only
required for SPP.

The main idea of the fix is that execution of vhost start/stop is moved
to eth_dev_start() and eth_dev_stop() from probe/remove.

Note that just one process which called eth_dev_start() can own this
vhost device although the vhost device is shared among the primary
process and secondary processes after that. Once eth_dev_stop() is
called from the owner for releasing the device, it can be owned from
another process again. It is user's responsibility that several
processes don't use the vhost device at once.

Signed-off-by: Itsuro Oda <oda@valinux.co.jp>
Acked-by: Yasufumi Ogawa <yasufum.o@gmail.com>