shared: make sure vhost stopped before adding
There is a case a secondary process which owned a vhost PMD and was down
before releasing the the PMD. This update makes sure the device is
stopped so that it is available to do following operations, such as
rte_eth_dev_configure(). It is no-op if the device is stopped.

Signed-off-by: Itsuro Oda <oda@valinux.co.jp>
Acked-by: Yasufumi Ogawa <yasufum.o@gmail.com>