spp_nfv: skip adding vhost PMD to port lists
ports_fwd_array and port_map are structs of spp_nfv which are lists of
ports possessed by the process. So a vhost device have to be owned by
just one process, it should not be added to the lists while initializing
if it is given with vdev option and already owned by another process.

Signed-off-by: Itsuro Oda <oda@valinux.co.jp>
Acked-by: Yasufumi Ogawa <yasufum.o@gmail.com>