shared/sec: skip adding vhosts to port list
In SPP worker processes, vhost device is managed in a list implemented
as a member of struct `p_iface_info`. However, it is not added to under
the management if it is already owned by another process. This update is
to exclude vhost is added to the list if it is already owned by another
process.

Signed-off-by: Itsuro Oda <oda@valinux.co.jp>
Acked-by: Yasufumi Ogawa <yasufum.o@gmail.com>