shared: revise vdev prefixs of pcap and nullpmd
In DPDK, there are two prefixes of RTE eth device, 'eth_' and 'net_' fo
r a reason of compatibility. 'eth_' is old one. SPP defines all of
prefixes as following.

  #define VDEV_ETH_VHOST "eth_vhost"
  #define VDEV_NET_VHOST "net_vhost"
  ...

However, pcap and nullpmd have only 'eth_' or 'net_'. SPP defines them
such as VDEV_NULL for simplicity, but it might be confusing for
considering the naming rule. This update is revise the prefixes to avoid
this ambiguity.

Signed-off-by: Yasufumi Ogawa <yasufum.o@gmail.com>