cli: fix error in asking ports if no pri forwarder
SPP CLI retrieves status of spp_primary not only for `status` command
but also others, such as `add` or `del`, to check the result of the
commands. For the commands other than `status`, SPP CLI is terminated
if it retrieves the status and forwarder thread is not running because
SPP CLI expects to receive the status including 'forwarder' attribute,
but does not exist.

This update is to fix the issue. SPP CLI tells you that forwarder thread
is not running for the cases.

Signed-off-by: Yasufumi Ogawa <yasufum.o@gmail.com>