cli: move logfile to under project log dir
SPP CLI outputs log as `spp/cli/log/spp_cli.log` because the contents
are mainly about implementation and no need for users. Some error
messages for users are just displayed in the CLI. However, it is
annoying that users have to remove the error message for failing
completion.

SPP CLI has got stabled and almost no need to output tiny logs of
implementation. It is the time to move CLI's log to `spp/log` as similar
to other processes.

This update is to move logfile to `spp/log/spp_cli.log`, and to show
the error messages in the logfile.

Signed-off-by: Yasufumi Ogawa <yasufum.o@gmail.com>