makefile: update dist-clean task
This patch is to add drivers to dist-clean task. In this update,
dist-clean is divided into two tasks. It is because files named as
`_postclean` is generated in `src/drivers/*/` directory after cleaning
and still remained. First task named as `_dist-clean` is for removing
files generated while compilation, and second one is for removing
`_postclean`.

Signed-off-by: Yasufumi Ogawa <yasufum.o@gmail.com>