tools/sppc: refactor container image names
A set of base names of container image is defined in `conf/env.py`
considering user uses any of container image name. However, build script
was changed to create image with default name and it is modified via
options. It means that user doesn't need to define the name anymore.
This update is to move the definition of base names from conf to common
because it is no need to edit usually.

Signed-off-by: Yasufumi Ogawa <yasufum.o@gmail.com>