tools/sppc: version checker for container DPDK ver
You need to check version of DPDK on a container because some of
application might be deprecated or removed from some specific version.
For instance, load balancer example app was removed after v19.08-rc.

This update is to add methods for getting DPDK version from container,
and comparing it with expected version.

Signed-off-by: Yasufumi Ogawa <yasufum.o@gmail.com>