tools/sppc: setup with docker opts in SPP pri
spp_primary launcher only does not use setup_docker_opts() because some
options are different from others, but almost similar actually. This
update is to merge primary's setup process into the method.

As a refactoring, name of container image is excluded from the result of
the method because it is not a option, but a mandatory param.

Signed-off-by: Yasufumi Ogawa <yasufum.o@gmail.com>